pistache: fix package config dependencies

+++ meta-oe/recipes-devtools/pistache/pistache_0.4.26.bb
@@ -9,22 +9,24 @@ SRCREV = "ddffda861aa49012dcda28f1362d0339e718cd52"
 
 S = "${WORKDIR}/git"
 
-DEPENDS = "openssl brotli zlib zstd rapidjson"
+DEPENDS = "rapidjson"
 
 inherit pkgconfig meson
 
 PACKAGECONFIG ?= "openssl brotli deflate zstd examples"
-PACKAGECONFIG[openssl]  = "-DPISTACHE_USE_SSL=true, -DPISTACHE_USE_SSL=false"
-PACKAGECONFIG[brotli]   = "-DPISTACHE_USE_CONTENT_ENCODING_BROTLI=true, -DPISTACHE_USE_CONTENT_ENCODING_BROTLI=false"
-PACKAGECONFIG[deflate]  = "-DPISTACHE_USE_CONTENT_ENCODING_DEFLATE=true, -DPISTACHE_USE_CONTENT_ENCODING_DEFLATE=false"
-PACKAGECONFIG[zstd]     = "-DPISTACHE_USE_CONTENT_ENCODING_ZSTD=true, -DPISTACHE_USE_CONTENT_ENCODING_ZSTD=false"
+PACKAGECONFIG[openssl]  = "-DPISTACHE_USE_SSL=true, -DPISTACHE_USE_SSL=false, openssl"
+PACKAGECONFIG[brotli]   = "-DPISTACHE_USE_CONTENT_ENCODING_BROTLI=true, -DPISTACHE_USE_CONTENT_ENCODING_BROTLI=false, brotli"
+PACKAGECONFIG[deflate]  = "-DPISTACHE_USE_CONTENT_ENCODING_DEFLATE=true, -DPISTACHE_USE_CONTENT_ENCODING_DEFLATE=false, zlib"
+PACKAGECONFIG[zstd]     = "-DPISTACHE_USE_CONTENT_ENCODING_ZSTD=true, -DPISTACHE_USE_CONTENT_ENCODING_ZSTD=false, zstd"
 PACKAGECONFIG[examples] = "-DPISTACHE_BUILD_EXAMPLES=true, -DPISTACHE_BUILD_EXAMPLES=false"
-PACKAGECONFIG[test]     = "-DPISTACHE_BUILD_TESTS=true, -DPISTACHE_BUILD_TESTS=false"
+PACKAGECONFIG[test]     = "-DPISTACHE_BUILD_TESTS=true, -DPISTACHE_BUILD_TESTS=false, gtest curl"
 
 do_install:append () {
-    install -d ${D}${bindir}
-    rm -rf ${B}/examples/*.p/
-    install -m 0755 ${B}/examples/run*   ${D}${bindir}/
+    if ${@bb.utils.contains("PACKAGECONFIG", "examples", "true", "false", d)}; then
+        install -d ${D}${bindir}
+        rm -rf ${B}/examples/*.p/
+        install -m 0755 ${B}/examples/run*   ${D}${bindir}/
+    fi
 }
 
 PACKAGES += "${PN}-examples"
